pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mudrex-signal-bot"
version = "1.0.0"
description = "Telegram Signal Bot for Mudrex Futures Trading"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "DecentralizedJM"}
]
keywords = ["mudrex", "telegram", "trading", "signals", "crypto", "futures"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "python-telegram-bot>=20.0",
    "mudrex-api-trading-python-sdk>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
mudrex-signal-bot = "signal_bot.run:main"

[tool.setuptools.packages.find]
where = ["src"]
